feat:#220 metadata를 layout으로 분리
